archive/backend_django/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.7
curl_cffi==0.6.0b7
dataclasses-json==0.6.3
Deprecated==1.2.14
distro==1.9.0
Django==5.0.1
django-cors-headers==4.3.1
django-oauth-toolkit==2.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
docstring-inheritance==2.1.2
drf-yasg==1.21.7
duckduckgo_search==4.4
exceptiongroup==1.2.0
frozenlist==1.4.1
future==1.0.0
gunicorn==21.2.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.6
inflection==0.5.1
jsonpatch==1.33
jsonpointer==2.4
jwcrypto==1.5.1
langchain==0.1.7
langchain-community==0.0.20
langchain-core==0.1.23
langchain-openai==0.0.6
langsmith==0.0.87
lxml==5.1.0
marshmallow==3.20.2
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.3
oauthlib==3.2.2
openai==1.10.0
packaging==23.2
pgvector==0.2.4
psycopg==3.1.17
psycopg2-binary==2.9.9
pycparser==2.21
pycryptodomex==3.20.0
pydantic==2.5.3
pydantic_core==2.14.6
pyjwkest==1.4.2
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.25
sqlparse==0.4.4
tenacity==8.2.3
tiktoken==0.5.2
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.1.0
wikipedia==1.4.0
wrapt==1.16.0
yarl==1.9.4
youtube-search==2.1.2
pypdf == 4.0.2
rapidocr_onnxruntime == 1.3.11
